Add files via upload
I tried to implement a small application with an Object-relational structural pattern. This application is responsible for reading data from the input and create objects regarding the data. Then it solves this problem of which teams and positions contain the valuable players(players with a value season of higher than 10).